requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioshutil==1.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
executing==2.2.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
griffe==1.13.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
litellm==1.76.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.13.1
mdurl==0.1.2
multidict==6.6.4
nest-asyncio==1.6.0
openai==1.102.0
openai-agents==0.2.9
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-zerox==0.0.7
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
rpds-py==0.27.0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
tiktoken==0.11.0
tokenizers==0.21.4
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
